Add minor edits

+++ frontend/components/users/user_stories_list.jsx
@@ -19,22 +19,44 @@ class UserStoriesList extends React.Component {
 
 
     render() {
-        const { stories } = this.props;
+        const { stories, user } = this.props;
         // console.log(stories)
         // let b = Object.entries(stories)
         // console.log(b)
-        const userStoryList = Object.entries(stories).map((story) => (
-            <li key={story[0]}>
-                <Link to={`/stories/${story[0]}`}>{story[1].title}</Link>
-                {/* <Link to={`/stories/${story.id}`}>{story.title}</Link> */}
-                <div className="user-story-btn">
-                    <Link to={`/stories/${story[1].id}/edit`}>Edit Story</Link>
-                    <button onClick={() => this.handleDelete(story[1].id)}>
-                        Delete Story
+        // console.log("this is username!!!!!!!!!!!")
+        // console.log(user.username)
+
+        // const userStoryList = Object.entries(stories).map((story) => (
+        //     <li key={story[0]}>
+        //         <Link to={`/stories/${story[0]}`}>{story[1].title}</Link>
+        //         {/* <Link to={`/stories/${story.id}`}>{story.title}</Link> */}
+        //         <div className="user-story-btn">
+        //             <Link to={`/stories/${story[1].id}/edit`}>Edit Story</Link>
+        //             <button onClick={() => this.handleDelete(story[1].id)}>
+        //                 Delete Story
+        //             </button>
+        //         </div>
+        //     </li>                                                                               
+        // ));
+
+        const userStoryList = Object.entries(stories).map((story) => {
+            if (story[1].author === user.username) {
+                return (
+                    <li key={story[0]}>
+                        <Link to={`/stories/${story[0]}`}>{story[1].title}</Link>
+                        {/* <Link to={`/stories/${story.id}`}>{story.title}</Link> */}
+                        <div className="user-story-btn">
+                            <Link to={`/stories/${story[1].id}/edit`}>Edit Story</Link>
+                            <button onClick={() => this.handleDelete(story[1].id)}>
+                                Delete Story
                     </button>
-                </div>
-            </li>                                                                               
-        ));
+                        </div>
+                    </li>   
+                )
+            } else {
+                null
+            }
+        })
 
         return (
             <div>
